Refactor predictor into separate modules
