+++ predictor.py
@@ -1,0 +1,586 @@
+import pandas as pd
+import numpy as np
+import fastf1
+from sklearn.metrics import mean_absolute_error
+
+from export_race_details import export_race_details
+from estimate_overtakes import average_overtakes
+from data_utils import (
+    _load_historical_data,
+    _clean_historical_data,
+    _add_driver_team_info,
+    _engineer_features,
+    _encode_features,
+    _load_overtake_stats,
+    _get_event_drivers,
+    _get_qualifying_results,
+    _get_fp3_results,
+    fetch_weather,
+    CIRCUIT_METADATA,
+    GRAND_PRIX_LIST,
+)
+from model_utils import _train_model, _rank_metrics
+
+try:
+    import shap  # type: ignore
+except Exception:
+    shap = None
+
+
+class SeasonSplit:
+    """Cross-validator that yields whole seasons as validation folds."""
+
+    def __init__(self, seasons, season_col="Season"):
+        self.seasons = list(seasons)
+        self.season_col = season_col
+
+    def split(self, X):
+        seasons_sorted = [s for s in self.seasons if s in X[self.season_col].unique()]
+        seasons_sorted.sort()
+        for i in range(1, len(seasons_sorted)):
+            train_idx = X.index[X[self.season_col].isin(seasons_sorted[:i])].to_numpy()
+            val_idx = X.index[X[self.season_col] == seasons_sorted[i]].to_numpy()
+            if len(train_idx) and len(val_idx):
+                yield train_idx, val_idx
+
+    def get_n_splits(self, X=None, y=None, groups=None):
+        return max(0, len(self.seasons) - 1)
+
+
+def predict_race(grand_prix, year=2025, export_details=False, debug=False, compute_overtakes=True):
+    schedule = fastf1.get_event_schedule(year)
+    match = schedule[schedule["EventName"].str.contains(grand_prix, case=False, na=False)]
+    if match.empty:
+        raise ValueError(f"Grand Prix '{grand_prix}' not found for {year}")
+    this_race_number = int(match.iloc[0]["RoundNumber"])
+    event_month = pd.to_datetime(match.iloc[0].get("EventDate"), errors="coerce").month
+
+    seasons = list(range(2020, year + 1))
+
+    overtake_map = _load_overtake_stats()
+    if compute_overtakes:
+        try:
+            years_for_avg = list(range(max(2020, year - 3), year))
+            avg = average_overtakes(grand_prix, years_for_avg)
+            overtake_map[grand_prix] = avg
+        except Exception as err:
+            print(f"⚠️ Could not compute overtakes for {grand_prix}: {err}")
+
+    race_data = _load_historical_data(seasons, overtake_map)
+    race_data = _clean_historical_data(race_data).reset_index(drop=True)
+    race_data['DriverNumber'] = pd.to_numeric(race_data['DriverNumber'], errors='coerce')
+    qual_results = None
+    fp3_results = None
+    race_data = _add_driver_team_info(race_data, seasons)
+    race_data = race_data.drop(columns=['Team'], errors='ignore')
+    race_data = race_data.loc[
+        ~((race_data['Season'] == year) & (race_data['RaceNumber'] >= this_race_number))
+    ].reset_index(drop=True)
+    race_data = _engineer_features(race_data)
+    race_data.sort_values(["Season", "RaceNumber"], inplace=True)
+    race_data.reset_index(drop=True, inplace=True)
+    race_data.to_csv("prediction_data.csv", index=False)
+
+    race_cols = [
+        'GridPosition', 'Season', 'ExperienceCount', 'IsRookie', 'TeamAvgPosition',
+        'CrossAvgFinish', 'RecentAvgPoints', 'BestQualiTime', 'MissedQuali',
+        'QualiPosition', 'FP3BestTime', 'FP3LongRunTime', 'DeltaToBestQuali',
+        'DeltaToNext', 'SprintFinish',
+        'Recent3AvgFinish', 'Recent5AvgFinish', 'DriverAvgTrackFinish',
+        'DriverTrackPodiums', 'DriverTrackDNFs', 'TeamRecentQuali',
+        'TeamRecentFinish', 'TeamReliability',
+        'DriverChampPoints', 'ConstructorChampPoints',
+        'DriverStanding', 'ConstructorStanding', 'PrevYearConstructorRank',
+        'CircuitLength', 'NumCorners', 'DRSZones', 'StdLapTime',
+        'IsStreet', 'DownforceLevel',
+        'AirTemp', 'TrackTemp', 'Rainfall', 'RainfallMissing', 'WeightedAvgOvertakes'
+    ]
+
+    holdout_year = year - 1
+    holdout_df = race_data[race_data['Season'] == holdout_year]
+    train_df = race_data[race_data['Season'] < holdout_year]
+    holdout_mae = None
+    holdout_rank = None
+    if not holdout_df.empty and not train_df.empty:
+        (
+            ho_feat,
+            ho_team_enc,
+            ho_circ_enc,
+            ho_top_circuits,
+            ho_top_teams,
+        ) = _encode_features(train_df, race_cols)
+        ho_val_feat, _, _, _, _ = _encode_features(
+            holdout_df,
+            race_cols,
+            ho_team_enc,
+            ho_circ_enc,
+            ho_top_circuits,
+            ho_top_teams,
+        )
+        ho_cv = SeasonSplit(sorted(train_df['Season'].unique()))
+        ho_model, _ = _train_model(ho_feat, train_df['Position'], ho_cv, debug)
+        ho_preds = ho_model.predict(ho_val_feat)
+        holdout_mae = mean_absolute_error(holdout_df['Position'], ho_preds)
+        holdout_rank = _rank_metrics(holdout_df['Position'], ho_preds)
+    (
+        features,
+        team_enc,
+        circuit_enc,
+        top_circuits,
+        top_teams,
+    ) = _encode_features(race_data, race_cols)
+    features = features.loc[race_data.index].reset_index(drop=True)
+
+    cv = SeasonSplit(sorted(race_data['Season'].unique()))
+
+    target = race_data['Position']
+    model, cv_rho = _train_model(features, target, cv, debug)
+
+    finish_preds_hist = model.predict(features)
+    finish_mae = mean_absolute_error(race_data['Position'], finish_preds_hist)
+    train_rank = _rank_metrics(race_data['Position'], finish_preds_hist)
+
+    try:
+        drivers_df = _get_qualifying_results(year, grand_prix)
+        drivers_df = drivers_df[drivers_df['BestTime'].notna()]
+        qual_results = drivers_df.copy()
+    except Exception:
+        drivers_df = _get_event_drivers(year, grand_prix)
+        qual_results = None
+
+    drivers_df['DriverNumber'] = pd.to_numeric(drivers_df['DriverNumber'], errors='coerce')
+
+    try:
+        fp3_results = _get_fp3_results(year, grand_prix)
+        if qual_results is not None:
+            qual_results = qual_results.merge(fp3_results, on='Abbreviation', how='left')
+    except Exception:
+        fp3_results = None
+
+    if qual_results is not None and not qual_results.empty:
+        default_best_q = qual_results['BestTime'].median()
+        default_delta_next = qual_results['DeltaToNext'].mean()
+    else:
+        default_best_q = race_data['BestQualiTime'].median()
+        default_delta_next = race_data['DeltaToNext'].mean()
+
+    if fp3_results is not None and not fp3_results.empty:
+        default_air = fp3_results['AvgAirTemp'].mean()
+        default_track = fp3_results['AvgTrackTemp'].mean()
+        default_rain = fp3_results['MaxRainfall'].max()
+        default_fp3 = fp3_results['FP3BestTime'].mean()
+        default_fp3_long = fp3_results['FP3LongRunTime'].mean()
+    else:
+        forecast = fetch_weather(grand_prix)
+        hist_air = race_data['AirTemp'].mean()
+        hist_track = race_data['TrackTemp'].mean()
+        hist_rain = race_data['Rainfall'].median()
+        default_fp3 = race_data['FP3BestTime'].mean()
+        default_fp3_long = race_data['FP3LongRunTime'].mean()
+        if forecast:
+            f_air = forecast['ForecastAirTemp']
+            f_track = f_air + 10
+            f_rain = forecast['ForecastPrecipChance']
+            default_air = (f_air + hist_air) / 2
+            default_track = (f_track + hist_track) / 2
+            default_rain = (f_rain + hist_rain) / 2
+        else:
+            default_air = hist_air
+            default_track = hist_track
+            default_rain = hist_rain
+    default_overtake = race_data['WeightedAvgOvertakes'].mean()
+
+    try:
+        from fastf1.circuit_info import get_circuit_info
+        circuit_lengths = {}
+        corners_map = {}
+        drs_map = {}
+        lap_map = {}
+        info = get_circuit_info(grand_prix)
+        length = (
+            info.get('Length')
+            or info.get('CircuitLength')
+            or info.get('circuitLength')
+        )
+        corners = (
+            info.get('NumberOfTurns')
+            or info.get('Turns')
+            or info.get('numCorners')
+        )
+        drs = (
+            info.get('NumberOfDRSZones')
+            or info.get('DRSZonesCount')
+            or info.get('drsZones')
+        )
+        laptime = (
+            info.get('LapTimeAvg')
+            or info.get('LapRecord')
+            or info.get('lapRecord')
+        )
+        if isinstance(length, str):
+            length = str(length).replace(' km', '')
+        if isinstance(laptime, str):
+            try:
+                laptime = pd.to_timedelta(laptime).total_seconds()
+            except Exception:
+                laptime = pd.to_numeric(laptime, errors='coerce')
+        circuit_lengths[grand_prix] = pd.to_numeric(length, errors='coerce')
+        corners_map[grand_prix] = pd.to_numeric(corners, errors='coerce')
+        drs_map[grand_prix] = pd.to_numeric(drs, errors='coerce')
+        lap_map[grand_prix] = pd.to_numeric(laptime, errors='coerce')
+    except Exception:
+        meta = CIRCUIT_METADATA.get(grand_prix, {})
+        circuit_lengths = {grand_prix: np.nan}
+        corners_map = {grand_prix: meta.get('NumCorners', np.nan)}
+        drs_map = {grand_prix: meta.get('DRSZones', np.nan)}
+        lap_map = {grand_prix: meta.get('StdLapTime', np.nan)}
+
+    TRACK_TYPE = {
+        'Monaco Grand Prix': 'street',
+        'Singapore Grand Prix': 'street',
+        'Las Vegas Grand Prix': 'street',
+    }
+    DOWNFORCE = {
+        'Monaco Grand Prix': 'high',
+        'Hungarian Grand Prix': 'high',
+        'Italian Grand Prix': 'low',
+        'Belgian Grand Prix': 'low',
+    }
+    df_level_map = {'low': 0, 'medium': 1, 'high': 2}
+    track_type_val = TRACK_TYPE.get(grand_prix, 'permanent')
+    is_street_val = 1 if track_type_val == 'street' else 0
+    downforce_val = DOWNFORCE.get(grand_prix, 'medium')
+    downforce_level_val = df_level_map.get(downforce_val, 1)
+    circuit_length_val = circuit_lengths.get(grand_prix, np.nan)
+    num_corners_val = corners_map.get(grand_prix, np.nan)
+    drs_zones_val = drs_map.get(grand_prix, np.nan)
+    std_lap_time_val = lap_map.get(grand_prix, np.nan)
+
+    pred_rows = []
+    default_team_avg = race_data['Position'].mean()
+    default_team_quali = race_data['QualiPosition'].mean()
+
+    driver_stats_lookup = race_data.set_index(['DriverNumber', 'Circuit'])[
+        ['DriverAvgTrackFinish', 'DriverTrackPodiums', 'DriverTrackDNFs']
+    ]
+
+    year_data = race_data[race_data['Season'] == year]
+    if not year_data.empty:
+        driver_pts = year_data.groupby('DriverNumber')['Points'].sum()
+        constructor_pts = year_data.groupby('HistoricalTeam')['Points'].sum()
+        driver_standings = driver_pts.rank(method='dense', ascending=False).astype(int)
+        constructor_standings = constructor_pts.rank(method='dense', ascending=False).astype(int)
+        driver_pts_map = driver_pts.to_dict()
+        constructor_pts_map = constructor_pts.to_dict()
+        driver_stand_map = driver_standings.to_dict()
+        constructor_stand_map = constructor_standings.to_dict()
+    else:
+        driver_pts_map = {}
+        constructor_pts_map = {}
+        driver_stand_map = {}
+        constructor_stand_map = {}
+
+    same_season_teams = race_data[race_data['Season'] == year]
+    if not same_season_teams.empty:
+        team_strength = (
+            same_season_teams.groupby('HistoricalTeam')['Position']
+            .mean()
+            .to_dict()
+        )
+    else:
+        team_strength = {}
+
+    prev_year = year - 1
+    prev_data = race_data[race_data['Season'] == prev_year]
+    if not prev_data.empty:
+        final_pts = prev_data.groupby('HistoricalTeam')['ConstructorChampPoints'].max()
+        prev_rank = final_pts.rank(method='dense', ascending=False)
+        prev_rank_map = prev_rank.to_dict()
+        default_prev_rank = int(prev_rank.max())
+    else:
+        prev_rank_map = {}
+        default_prev_rank = 0
+
+    if qual_results is not None and not qual_results.empty:
+        fastest = qual_results['BestTime'].min()
+        qual_results['DeltaToBestQuali'] = qual_results['BestTime'] - fastest
+        qual_results = qual_results.sort_values('BestTime')
+        qual_results['DeltaToNext'] = qual_results['BestTime'].diff(-1).abs()
+        qual_results = qual_results.sort_index()
+        team_mean = qual_results.groupby('Team')['BestTime'].transform('mean')
+        team_size = qual_results.groupby('Team')['BestTime'].transform('size')
+        qual_results['DeltaToTeammateQuali'] = np.where(team_size > 1,
+                                                       (qual_results['BestTime'] - team_mean) * 2,
+                                                       0)
+        if 'Q1' in qual_results.columns and 'Q3' in qual_results.columns:
+            qual_results['QualiSessionGain'] = qual_results['Q1'] - qual_results['Q3']
+            std = qual_results['QualiSessionGain'].std()
+            qual_results['QualiSessionGain'] = (
+                (qual_results['QualiSessionGain'] - qual_results['QualiSessionGain'].mean()) / std
+            ) if std != 0 else 0
+        else:
+            qual_results['QualiSessionGain'] = 0
+        qual_results['GridDropCount'] = 0
+
+    driver_iter = qual_results if qual_results is not None and not qual_results.empty else drivers_df
+    overall_avg_pos = race_data['Position'].mean()
+    rookie_avg_pos = race_data[race_data['ExperienceCount'] == 1]['Position'].mean()
+    for _, d in driver_iter.iterrows():
+        exp_count = len(race_data[race_data['DriverNumber'] == d['DriverNumber']]) + 1
+        team_name = d['Team']
+        team_same_season = race_data[
+            (race_data['HistoricalTeam'] == team_name) &
+            (race_data['Season'] == year)
+        ].sort_values('RaceNumber')
+        team_prev = team_same_season[team_same_season['RaceNumber'] < this_race_number]
+
+        if len(team_prev) > 0:
+            team_avg_pos = team_prev['Position'].mean()
+        else:
+            team_avg_pos = team_strength.get(team_name, default_team_avg)
+
+        team_prev_q = team_prev['QualiPosition']
+        if len(team_prev_q) > 0:
+            team_recent_q = team_prev_q.tail(5).mean()
+        else:
+            team_recent_q = default_team_quali
+
+        team_prev_f = team_prev['Position']
+        if len(team_prev_f) > 0:
+            team_recent_f = team_prev_f.tail(5).mean()
+        else:
+            team_recent_f = default_team_avg
+
+        team_prev_dnf = team_prev['DidNotFinish']
+        if len(team_prev_dnf) > 0:
+            team_rel = team_prev_dnf.tail(5).sum()
+        else:
+            team_rel = 0.0
+
+        stats = driver_stats_lookup.loc[(d['DriverNumber'], grand_prix)] if (d['DriverNumber'], grand_prix) in driver_stats_lookup.index else None
+        if stats is None:
+            avg_track = race_data['DriverAvgTrackFinish'].mean()
+            podiums = 0.0
+            dnfs = 0.0
+        else:
+            avg_track = stats['DriverAvgTrackFinish']
+            podiums = stats['DriverTrackPodiums']
+            dnfs = stats['DriverTrackDNFs']
+
+        if qual_results is not None and 'GridPosition' in d and pd.notna(d['GridPosition']):
+            grid_pos = int(d['GridPosition'])
+            best_time = d['BestTime']
+        else:
+            grid_pos = 20
+            best_time = default_best_q
+
+        if fp3_results is not None and 'FP3BestTime' in d and pd.notna(d['FP3BestTime']):
+            fp3_time = d['FP3BestTime']
+        else:
+            fp3_time = default_fp3
+
+        if fp3_results is not None and 'FP3LongRunTime' in d and pd.notna(d['FP3LongRunTime']):
+            fp3_long_time = d['FP3LongRunTime']
+        else:
+            fp3_long_time = default_fp3_long
+
+        driver_num = int(d['DriverNumber'])
+        past_races = race_data[
+            (race_data['DriverNumber'] == driver_num) & (
+                (race_data['Season'] < year) |
+                ((race_data['Season'] == year) & (race_data['RaceNumber'] < this_race_number))
+            )
+        ].sort_values(['Season', 'RaceNumber'])
+
+        if past_races.empty:
+            cross_avg = rookie_avg_pos
+            recent_avg_pts = 0.0
+            recent3_avg = rookie_avg_pos
+            recent5_avg = rookie_avg_pos
+        else:
+            cross_avg = past_races['Position'].tail(5).mean()
+            recent_avg_pts = past_races['Points'].tail(5).mean()
+            recent3_avg = past_races['Position'].tail(3).mean()
+            recent5_avg = past_races['Position'].tail(5).mean()
+        pred_rows.append({
+            'GridPosition': grid_pos,
+            'Season': year,
+            'ExperienceCount': exp_count,
+            'IsRookie': 1 if exp_count == 1 else 0,
+            'TeamAvgPosition': team_avg_pos,
+            'CrossAvgFinish': cross_avg,
+            'RecentAvgPoints': recent_avg_pts,
+            'BestQualiTime': best_time,
+            'QualiPosition': grid_pos,
+            'FP3BestTime': fp3_time,
+            'FP3LongRunTime': fp3_long_time,
+            'DeltaToBestQuali': d.get('DeltaToBestQuali', 0),
+            'DeltaToNext': d.get('DeltaToNext', default_delta_next),
+            'SprintFinish': 25,
+            'Recent3AvgFinish': recent3_avg,
+            'Recent5AvgFinish': recent5_avg,
+            'DriverAvgTrackFinish': avg_track,
+            'DriverTrackPodiums': podiums,
+            'DriverTrackDNFs': dnfs,
+            'TeamRecentQuali': team_recent_q,
+            'TeamRecentFinish': team_recent_f,
+            'TeamReliability': team_rel,
+            'DriverChampPoints': driver_pts_map.get(d['DriverNumber'], 0.0),
+            'ConstructorChampPoints': constructor_pts_map.get(d['Team'], 0.0),
+            'DriverStanding': int(driver_stand_map.get(d['DriverNumber'], 0)),
+            'ConstructorStanding': int(constructor_stand_map.get(d['Team'], 0)),
+            'PrevYearConstructorRank': prev_rank_map.get(team_name, default_prev_rank),
+            'CircuitLength': circuit_length_val,
+            'NumCorners': num_corners_val,
+            'DRSZones': drs_zones_val,
+            'StdLapTime': std_lap_time_val,
+            'IsStreet': is_street_val,
+            'DownforceLevel': downforce_level_val,
+            'AirTemp': default_air,
+            'TrackTemp': default_track,
+            'Rainfall': default_rain,
+            'WeightedAvgOvertakes': default_overtake,
+            'Team': d['Team'],
+            'FullName': d['FullName'],
+            'Abbreviation': d['Abbreviation']
+        })
+
+    pred_df = pd.DataFrame(pred_rows)
+
+    if pred_df.empty:
+        raise ValueError(f"No driver data available for {year} {grand_prix}")
+
+    pred_df['GridPosition'] = (
+        pd.to_numeric(pred_df['GridPosition'], errors='coerce')
+        .fillna(20)
+    )
+    pred_df['GridPosition'] = pred_df['GridPosition'].clip(1, 20)
+    pred_df['QualiPosition'] = (
+        pd.to_numeric(pred_df['QualiPosition'], errors='coerce')
+        .fillna(20)
+    )
+    pred_df['BestQualiTime'] = pd.to_numeric(pred_df['BestQualiTime'], errors='coerce')
+    pred_df['MissedQuali'] = pred_df['BestQualiTime'].isna().astype(int)
+    pred_df['BestQualiTime'] = pred_df['BestQualiTime'].fillna(
+        race_data['BestQualiTime'].median()
+    )
+    pred_df['FP3BestTime'] = (
+        pd.to_numeric(pred_df['FP3BestTime'], errors='coerce')
+        .fillna(race_data['FP3BestTime'].mean())
+    )
+    pred_df['FP3LongRunTime'] = (
+        pd.to_numeric(pred_df['FP3LongRunTime'], errors='coerce')
+        .fillna(race_data['FP3LongRunTime'].mean())
+    )
+    pred_df['SprintFinish'] = (
+        pd.to_numeric(pred_df.get('SprintFinish'), errors='coerce')
+        .fillna(25)
+    )
+    pred_df['NumCorners'] = (
+        pd.to_numeric(pred_df['NumCorners'], errors='coerce')
+        .fillna(race_data['NumCorners'].median())
+    )
+    pred_df['DRSZones'] = (
+        pd.to_numeric(pred_df['DRSZones'], errors='coerce')
+        .fillna(race_data['DRSZones'].median())
+    )
+    pred_df['StdLapTime'] = (
+        pd.to_numeric(pred_df['StdLapTime'], errors='coerce')
+        .fillna(race_data['StdLapTime'].mean())
+    )
+
+    pred_df['Month'] = event_month
+
+    air_map = race_data.groupby(['Circuit', 'Month'])['AirTemp'].median()
+    air_val = air_map.get((grand_prix, event_month), np.nan)
+    pred_df['AirTemp'] = pd.to_numeric(pred_df['AirTemp'], errors='coerce')
+    pred_df['AirTemp'] = pred_df['AirTemp'].fillna(air_val)
+    pred_df['AirTemp'] = pred_df['AirTemp'].fillna(race_data['AirTemp'].mean())
+
+    track_map = race_data.groupby(['Circuit', 'Month'])['TrackTemp'].median()
+    track_val = track_map.get((grand_prix, event_month), np.nan)
+    pred_df['TrackTemp'] = pd.to_numeric(pred_df['TrackTemp'], errors='coerce')
+    pred_df['TrackTemp'] = pred_df['TrackTemp'].fillna(track_val)
+    pred_df['TrackTemp'] = pred_df['TrackTemp'].fillna(race_data['TrackTemp'].mean())
+
+    pred_df['Rainfall'] = pd.to_numeric(pred_df['Rainfall'], errors='coerce')
+    pred_df['RainfallMissing'] = pred_df['Rainfall'].isna().astype(int)
+    rain_map = race_data.groupby(['Circuit', 'Month'])['Rainfall'].median()
+    rain_val = rain_map.get((grand_prix, event_month), np.nan)
+    pred_df['Rainfall'] = pred_df['Rainfall'].fillna(rain_val)
+    circ_rain_map = race_data.groupby('Circuit')['Rainfall'].median()
+    circuit_rain = circ_rain_map.get(grand_prix, race_data['Rainfall'].median())
+    pred_df['Rainfall'] = pred_df['Rainfall'].fillna(circuit_rain)
+    pred_df['Rainfall'] = pred_df['Rainfall'].fillna(race_data['Rainfall'].median())
+    pred_df = pred_df.drop(columns=['Month'], errors='ignore')
+
+    pred_df.to_csv("prediction_input.csv", index=False)
+
+    race_pred_features, _, _, _, _ = _encode_features(
+        pred_df.assign(Circuit=grand_prix),
+        race_cols,
+        team_enc,
+        circuit_enc,
+        top_circuits,
+        top_teams,
+    )
+
+    pred_features = race_pred_features.reindex(columns=features.columns, fill_value=0)
+
+    preds = model.predict(pred_features)
+    shap_values = None
+    if debug and shap is not None:
+        try:
+            explainer = shap.TreeExplainer(model)
+            shap_values = explainer.shap_values(pred_features)
+        except Exception as err:
+            print(f"⚠️ Could not compute SHAP values: {err}")
+            shap_values = None
+
+    results = pd.DataFrame({
+        'Driver': pred_df['Abbreviation'],
+        'Team': pred_df['Team'],
+        'Grid': pred_df['GridPosition'],
+        'Predicted_Position': preds
+    })
+    sort_idx = results['Predicted_Position'].argsort()
+    results = results.iloc[sort_idx].reset_index(drop=True)
+    if shap_values is not None:
+        shap_values = shap_values[sort_idx]
+    results['Final_Position'] = range(1, len(results) + 1)
+    results.to_csv("prediction_results.csv", index=False)
+    if export_details:
+        try:
+            detail_path = export_race_details(year, grand_prix)
+            print(f"📁 Saved session data to {detail_path}")
+        except Exception as err:
+            print(f"⚠️ Could not export session data: {err}")
+    details = None
+    if debug:
+        details = {
+            'prediction_features': pred_features,
+            'shap_values': shap_values,
+            'feature_names': pred_features.columns.tolist(),
+        }
+
+    if holdout_mae is not None:
+        print(
+            f"📊 CV Spearman: {cv_rho:.2f} -- Hold-out MAE: {holdout_mae:.2f} -- Training MAE: {finish_mae:.2f}"
+        )
+        print(
+            f"📈 Spearman \u03c1: {train_rank['spearman']:.2f} (train) / {holdout_rank['spearman']:.2f} (hold-out) "
+            f"-- Top1: {train_rank['top1']*100:.0f}% / {holdout_rank['top1']*100:.0f}% "
+            f"-- Top3: {train_rank['top3']*100:.0f}% / {holdout_rank['top3']*100:.0f}%"
+        )
+    else:
+        print(f"📊 CV Spearman: {cv_rho:.2f} -- Training MAE: {finish_mae:.2f}")
+        print(
+            f"📈 Spearman \u03c1: {train_rank['spearman']:.2f} -- "
+            f"Top1: {train_rank['top1']*100:.0f}% -- Top3: {train_rank['top3']*100:.0f}%"
+        )
+    return (results, details) if debug else results
+
+
+if __name__ == '__main__':
+    res = predict_race('Chinese Grand Prix', year=2025, export_details=True, debug=False)
+    print(res[['Driver', 'Team', 'Grid', 'Final_Position']].head())
